server/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.8
structlog>=24.1
python-dotenv>=1.0
cachetools>=5.3
types-cachetools>=6.1.0
mcp[cli]>=1.2
pytest>=8.2
pytest-cov>=4.1
ruff>=0.1.0
mypy>=1.6
redis>=4.6
orjson>=3.9
tenacity>=8.2
fastapi-limiter>=0.1
pydantic-settings>=2.0
types-redis>=4.6
